more progess on tabel
